Added license, broke classes into separate source files
