+++ dumpi/libotf2dump/otf2deftable.cc
@@ -1,0 +1,83 @@
+/**
+Copyright 2009-2017 National Technology and Engineering Solutions of Sandia,
+LLC (NTESS).  Under the terms of Contract DE-NA-0003525, the U.S.  Government
+retains certain rights in this software.
+
+Sandia National Laboratories is a multimission laboratory managed and operated
+by National Technology and Engineering Solutions of Sandia, LLC., a wholly
+owned subsidiary of Honeywell International, Inc., for the U.S. Department of
+Energy's National Nuclear Security Administration under contract DE-NA0003525.
+
+Copyright (c) 2009-2017, NTESS
+
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met:
+
+    * Redistributions of source code must retain the above copyright
+      notice, this list of conditions and the following disclaimer.
+
+    * Redistributions in binary form must reproduce the above
+      copyright notice, this list of conditions and the following
+      disclaimer in the documentation and/or other materials provided
+      with the distribution.
+
+    * Neither the name of Sandia Corporation nor the names of its
+      contributors may be used to endorse or promote products derived
+      from this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+Questions? Contact sst-macro-help@sandia.gov
+*/
+
+#include <dumpi/libotf2dump/otf2writer.h>
+#include <sys/stat.h>
+#include <algorithm>
+#include <limits.h>
+#include <cstring>
+#include <unordered_set>
+#include <functional>
+
+using std::string;
+using std::to_string;
+
+namespace dumpi {
+
+  OTF2DefTable::OTF2DefTable() {}
+  int OTF2DefTable::map(string string)
+  {
+    auto it = _map.find(string);
+    _added_last_lookup = false;
+    if ( it != _map.end()) return it->second;
+    _map.insert(std::make_pair(string, _counter));
+    _added_last_lookup = true;
+    return _counter++;
+  }
+
+  const string OTF2DefTable::map(int index) {
+    // Finds the string corresponding to the given int.
+    for (auto elem = _map.begin(); elem != _map.end(); elem++)
+      if (elem->second == index)
+        return elem->first;
+    return string();
+  }
+
+  const string OTF2DefTable::operator[] (int index) { return map(index); }
+  int OTF2DefTable::operator[] (const char* index) { return map(string(index));}
+  int OTF2DefTable::operator[] (string string) { return map(string); }
+  int OTF2DefTable::size() { return _counter; }
+  bool OTF2DefTable::added_last_lookup() {return _added_last_lookup;}
+
+}
